requirements.txt
kivymd==1.1.1
pyaml==23.7.0
sqlalchemy==2.0.20
igdb-api-v4==0.2.0
beautifulsoup4==4.12.2
pandas==2.2.3
requests==2.31.0
lxml==4.9.3
qbittorrent-api==2023.7.52
certifi